refactor Time tests
