--- test/helpers/map-values.js
@@ -1,7 +1,0 @@
-function mapValues(obj, fn) {
-  return Object.fromEntries(Object.entries(obj).map(([k, v]) => [k, fn(v)]));
-}
-
-module.exports = {
-  mapValues,
-};
